+++ vitest.ssr.config.ts
@@ -8,6 +8,7 @@ export default defineConfig({
       '@gluonjs/reactivity': resolve(import.meta.dirname, 'packages/reactivity/src/index.ts'),
       '@gluonjs/router/memory': resolve(import.meta.dirname, 'packages/router/src/memory.ts'),
       '@gluonjs/ssr/streaming': resolve(import.meta.dirname, 'packages/ssr/src/streaming.ts'),
+      '@gluonjs/ssr/static': resolve(import.meta.dirname, 'packages/ssr/src/static.ts'),
       '@gluonjs/ssr': resolve(import.meta.dirname, 'packages/ssr/src/index.ts'),
       '@gluonjs/store': resolve(import.meta.dirname, 'packages/store/src/index.ts'),
     },
@@ -17,7 +18,7 @@ export default defineConfig({
     include: ['tests-node/ssr.spec.ts'],
     coverage: {
       provider: 'v8',
-      include: ['packages/ssr/src/index.ts', 'packages/ssr/src/streaming.ts'],
+      include: ['packages/ssr/src/index.ts', 'packages/ssr/src/static.ts', 'packages/ssr/src/streaming.ts'],
       reportsDirectory: 'coverage/ssr',
       reporter: ['text', 'html'],
       thresholds: {